+++ lib/ftpd/file_system_error_translator.rb
@@ -1,7 +1,7 @@
 module Ftpd
 
-  # A proxy file system driver that sends "450" replies when the
-  # wrapped file system driver raises a FileSystemError.
+  # A proxy file system driver that sends a "450" or "550" error
+  # reply in response to FileSystemError exceptions.
 
   class FileSystemErrorTranslator
 
@@ -17,8 +17,12 @@ module Ftpd
 
     def method_missing(method, *args)
       @file_system.send(method, *args)
+    rescue PermanentFileSystemError => e
+      permanent_error e
+    rescue TransientFileSystemError => e
+      transient_error e
     rescue FileSystemError => e
-      error "450 #{e}"
+      permanent_error e
     end
 
   end